% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ACE.13.TCE.df}
\alias{ACE.13.TCE.df}
\title{Trichloroethylene Concentrations Before and After Remedation}
\usage{
data(ACE.13.TCE.df)
}
\description{
Trichloroethylene (TCE) concentrations (mg/L) at 10 groundwater monitoring wells
  before and after remediation.
}
\docType{data}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{TCE.mg.per.L}}{TCE concentrations}
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Period}}{a factor indicating the period (before vs. after remediation)}
  }
}
\source{
  USACE. (2013).  \emph{Environmental Quality - Environmental Statistics}.
  Engineer Manual EM 200-1-16, 31 May 2013.  Department of the Army,
  U.S. Army Corps of Engineers, Washington, D.C. 20314-1000, p. M-10.
  \url{https://www.publications.usace.army.mil/Portals/76/Publications/EngineerManuals/EM_200-1-16.pdf}.
}
\keyword{datasets}
