% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enpar.R
\name{enpar}
\alias{enpar}
\title{Estimate Mean, Standard Deviation, and Standard Error Nonparametrically}
\usage{
enpar(x, ci = FALSE, ci.method = "bootstrap", ci.type = "two-sided",
      conf.level = 0.95, pivot.statistic = "z", n.bootstraps = 1000, seed = NULL)
}
\description{
Estimate the mean, standard deviation, and standard error of the mean
  nonparametrically given a sample of data, and optionally construct
  a confidence interval for the mean.
}
\details{
This help page has been shortened to keep function help focused on usage,
arguments, return values, and examples. Extended method details, formulas,
and background material are available in \code{vignette("extended-function-details", package = "EnvStats")}, section \code{enpar}.
}
\arguments{
  \item{x}{
  numeric vector of observations.
  Missing (\code{NA}), undefined (\code{NaN}), and
  infinite (\code{Inf}, \code{-Inf}) values are allowed but will be removed.
}
  \item{ci}{
  logical scalar indicating whether to compute a confidence interval for the
  mean.  The default value is \code{ci=FALSE}.
}
  \item{ci.method}{
  character string indicating what method to use to construct the confidence interval
  for the mean.  The possible values are
  \code{"bootstrap"} (based on bootstrapping; the default), and
  \code{"normal.approx"} (normal approximation).
  See the \bold{DETAILS} section for more information.
  This argument is ignored if \code{ci=FALSE}.
}
  \item{ci.type}{
  character string indicating what kind of confidence interval to compute.  The
  possible values are \code{"two-sided"} (the default), \code{"lower"}, and
  \code{"upper"}.  This argument is ignored if \code{ci=FALSE}.
}
  \item{conf.level}{
  a scalar between 0 and 1 indicating the confidence level of the confidence interval.
  The default value is \code{conf.level=0.95}. This argument is ignored if
  \code{ci=FALSE}.
}
  \item{pivot.statistic}{
  character string indicating which statistic to use for the confidence interval
  for the mean when \code{ci.method="normal.approx"}.  Possible values are
  \code{"z"} (confidence interval based on the z-statistic; the default), and
  \code{"t"} (confidence interval based on the t-statistic).
  This argument is ignored if \code{ci=FALSE} or \cr
  \code{ci.method="bootstrap"}.
}
  \item{n.bootstraps}{
  numeric scalar indicating how many bootstraps to use to construct the
  confidence interval for the mean.  This argument is ignored if
  \code{ci=FALSE} or \cr
  \code{ci.method="normal.approx"}.
}
  \item{seed}{
  integer supplied to the function \code{\link[base]{set.seed}} and used when \cr
  \code{ci=TRUE} and \code{ci.method="bootstrap"}.  The default value is
  \code{seed=NULL}, in which case the current value of \code{.Random.seed}
  is used.  This argument is ignored if \code{ci=FALSE} or
  \code{ci.method="normal.approx"}.  This argument is necessary to create
  reproducible results for the bootstrapped confidence intervals
  (see the \bold{EXAMPLES} section).
}
}
\value{
  a list of class \code{"estimate"} containing the estimated parameters
  and other information.  See \code{\link{estimate.object}} for details.
}
\references{
  Efron, B. (1979).  Bootstrap Methods: Another Look at the Jackknife.
  \emph{The Annals of Statistics} \bold{7}, 1--26.

  Efron, B., and R.J. Tibshirani. (1993).  \emph{An Introduction to the Bootstrap}.
  Chapman and Hall, New York, 436pp.
}
\author{
    Steven P. Millard (\email{EnvStats@ProbStatInfo.com})
}
\note{
  The function \code{enpar} is related to the companion function
  \code{\link{enparCensored}} for censored data.  To estimate the median and
  compute a confidence interval, use \code{\link{eqnpar}}.

  The result of the call to \code{enpar} with \code{ci.method="normal.approx"}
  and \code{pivot.statistic="t"} produces the same result as the call to
  \code{\link{enorm}} with \code{ci.param="mean"}.
}
\seealso{
  \code{\link{enparCensored}}, \code{\link{eqnpar}}, \code{\link{enorm}},
  \code{\link{mean}}, \code{\link{sd}}, \code{\link{estimate.object}}.
}
\examples{
  # The data frame ACE.13.TCE.df contains observations on
  # Trichloroethylene (TCE) concentrations (mg/L) at
  # 10 groundwater monitoring wells before and after remediation.
  #
  # Compute the mean concentration for each period along with
  # a 95% bootstrap BCa confidence interval for the mean.
  #
  # NOTE: Use of the argument "seed" is necessary to reproduce this example.
  #
  # Before remediation: 21.6 [14.2, 30.1]
  # After remediation:   3.6 [ 1.6,  5.7]

  with(ACE.13.TCE.df,
    enpar(TCE.mg.per.L[Period=="Before"], ci = TRUE, seed = 476))

  #Results of Distribution Parameter Estimation
  #--------------------------------------------
  #
  #Assumed Distribution:            None
  #
  #Estimated Parameter(s):          mean    = 21.62400
  #                                 sd      = 13.51134
  #                                 se.mean =  4.27266
  #
  #Estimation Method:               Sample Mean
  #
  #Data:                            TCE.mg.per.L[Period == "Before"]
  #
  #Sample Size:                     10
  #
  #Confidence Interval for:         mean
  #
  #Confidence Interval Method:      Bootstrap
  #
  #Number of Bootstraps:            1000
  #
  #Confidence Interval Type:        two-sided
  #
  #Confidence Level:                95%
  #
  #Confidence Interval:             Pct.LCL = 13.95560
  #                                 Pct.UCL = 29.79510
  #                                 BCa.LCL = 14.16080
  #                                 BCa.UCL = 30.06848
  #                                 t.LCL   = 12.41945
  #                                 t.UCL   = 32.47306

  #----------

  with(ACE.13.TCE.df,
    enpar(TCE.mg.per.L[Period=="After"], ci = TRUE, seed = 543))

  #Results of Distribution Parameter Estimation
  #--------------------------------------------
  #
  #Assumed Distribution:            None
  #
  #Estimated Parameter(s):          mean    = 3.632900
  #                                 sd      = 3.554419
  #                                 se.mean = 1.124006
  #
  #Estimation Method:               Sample Mean
  #
  #Data:                            TCE.mg.per.L[Period == "After"]
  #
  #Sample Size:                     10
  #
  #Confidence Interval for:         mean
  #
  #Confidence Interval Method:      Bootstrap
  #
  #Number of Bootstraps:            1000
  #
  #Confidence Interval Type:        two-sided
  #
  #Confidence Level:                95%
  #
  #Confidence Interval:             Pct.LCL = 1.833843
  #                                 Pct.UCL = 5.830230
  #                                 BCa.LCL = 1.631655
  #                                 BCa.UCL = 5.677514
  #                                 t.LCL   = 1.683791
  #                                 t.UCL   = 8.101829
}
\keyword{ distribution }
\keyword{ htest }
