% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EnvStats-internal}
\alias{EnvStats-internal}
\alias{StatNText}
\alias{StatMeanSDText}
\alias{StatMedianIQRText}
\alias{StatTestText}
\title{Internal EnvStats Objects}
\description{
Internal EnvStats objects
}
\details{
  These are not to be called by the user.  They have been exported to allow advanced users to
  see their structure.
}
\keyword{ internal }
