% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.02d.Ex.2.ug.per.L.vec}
\alias{EPA.02d.Ex.2.ug.per.L.vec}
\title{Concentrations in Exhibit 2 of 2002d USEPA Guidance Document}
\usage{
data(EPA.02d.Ex.2.ug.per.L.vec)
}
\description{
Concentrations (\eqn{mu}g/L) from an exposure unit.
}
\docType{data}
\format{
  a numeric vector of concentrations (\eqn{mu}g/L)
}
\source{
  USEPA. (2002d).  \emph{Calculating Upper Confidence Limits for Exposure Point Concentrations at
  Hazardous Waste Sites}.  OSWER 9285.6-10, December 2002.  Office of Emergency and Remedial Response,
  U.S. Environmental Protection Agency, Washington, D.C., p. 9.
}
\keyword{datasets}
