% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Ex.12.1.ccl4.df}
\alias{EPA.09.Ex.12.1.ccl4.df}
\title{Carbon Tetrachloride Concentrations from Example 12-1 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.12.1.ccl4.df
}
\description{
Carbon tetrachloride (CCL4) concentrations (ppb) at five background wells (four measures at each well).
}
\docType{data}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{CCL4.ppb}}{a numeric vector of CCL4 concentrations (ppb)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.12-3.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
