% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Ex.14.4.arsenic.df}
\alias{EPA.09.Ex.14.4.arsenic.df}
\title{Arsenic Concentrations from Example 14-4 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.14.4.arsenic.df
}
\description{
Sixteen quarterly measures of arsenic concentrations (ppb).
}
\docType{data}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{Sample.Date}}{a factor indicating the month and year of collection}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Year}}{a factor indicating the year of collection}
    \item{\code{Arsenic.ppb}}{a numeric vector of arsenic concentrations (ppb)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.14-18.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
