% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Ex.14.8.df}
\alias{EPA.09.Ex.14.8.df}
\title{Analyte Concentrations from Example 14-8 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.14.8.df
}
\description{
Monthly unadjusted and adjusted analyte concentrations over a 3-year period.  Adjusted
  concentrations are computed by subtracting the monthly mean and adding the overall mean.
}
\docType{data}
\format{
  A data frame with 36 observations on the following 4 variables.
  \describe{
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Year}}{a numeric vector indicating the year of collection}
    \item{\code{Unadj.Conc}}{a numeric vector of unadjusted concentrations}
    \item{\code{Adj.Conc}}{a numeric vector adjusted concentrations}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.14-32.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
