% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.09.Ex.16.1.sulfate.df}
\alias{EPA.09.Ex.16.1.sulfate.df}
\title{Sulfate Concentrations from Example 16-1 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.16.1.sulfate.df
}
\description{
Sulfate concentrations (ppm) at one background well and one downgradient well
  (eight quarterly measures at each well).
}
\docType{data}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Year}}{a factor indicating the year of collection}
    \item{\code{Well.type}}{a factor indicating the well type (background vs. downgradient)}
    \item{\code{Sulfate.ppm}}{a numeric vector of sulfate concentrations (ppm)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.16-6.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
