% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Ex.17.7.sodium.df}
\alias{EPA.09.Ex.17.7.sodium.df}
\title{Sodium Concentrations from Example 17-7 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.17.7.sodium.df
}
\description{
Sodium concentrations (ppm) collected over several years.  The sample dates are
  recorded as the year of collection (2-digit format) plus a fractional part indicating when during the year the sample was collected.
}
\docType{data}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{Year}}{a numeric vector indicating the year of collection (a fractional number)}
    \item{\code{Sodium.ppm}}{a numeric vector of sodium concentrations (ppm)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.17-36.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
