% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Ex.19.1.sulfate.df}
\alias{EPA.09.Ex.19.1.sulfate.df}
\title{Sulfate Concentrations from Example 19-1 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.19.1.sulfate.df
}
\description{
Sulfate concentrations (mg/L) at four background wells.
}
\docType{data}
\format{
  A data frame with 25 observations on the following 7 variables.
  \describe{
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Month}}{a numeric vector indicating the month of collection}
    \item{\code{Day}}{a numeric vector indicating the day of the month of collection}
    \item{\code{Year}}{a numeric vector indicating the year of collection}
    \item{\code{Date}}{a Date object indicating the date of collection}
    \item{\code{Sulfate.mg.per.l}}{a numeric vector of sulfate concentrations (mg/L)}
    \item{\code{log.Sulfate.mg.per.l}}{a numeric vector of log-transformed sulfate concentrations (mg/L)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.19-17.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
