% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.09.Ex.7.1.arsenic.df}
\alias{EPA.09.Ex.7.1.arsenic.df}
\title{Arsenic concentrations from Example 7.1 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.7.1.arsenic.df
}
\description{
Arsenic concentrations (\eqn{\mu}g/L) at a single well, consisting of:
  8 historical observations,
  4 future observations for Case 1, and
  4 future observations for Case 2.
}
\docType{data}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{Data.Source}}{a factor with levels \code{Historical}, \code{Case.1}, \code{Case.2}}
    \item{\code{Arsenic.ug.per.l}}{a numeric vector of arsenic concentrations (\eqn{\mu}g/L)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.7-26.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
