% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Table.9.4.nickel.vec}
\alias{EPA.09.Table.9.4.nickel.vec}
\title{Nickel Concentrations in Table 9-4 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Table.9.4.nickel.vec
}
\description{
Nickel concentrations (ppb) from a single well.
}
\docType{data}
\format{
  a numeric vector of nickel concentrations (ppb)
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.9-18.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
