% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.89b.aldicarb1.df}
\alias{EPA.89b.aldicarb1.df}
\title{Aldicarb Concentrations from 1989 USEPA Guidance Document}
\usage{
EPA.89b.aldicarb1.df
}
\description{
Aldicarb concentrations (ppb) at three compliance wells (four monthly samples at each well).
}
\docType{data}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{Aldicarb}}{Aldicarb concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1989b). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities, Interim Final Guidance}.
  EPA/530-SW-89-026. Office of Solid Waste, U.S. Environmental Protection Agency,
  Washington, D.C. p.6-4.
}
\keyword{datasets}
