% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.89b.loglead.df}
\alias{EPA.89b.loglead.df}
\title{Log-transformed Lead Concentrations from 1989 USEPA Guidance Document}
\usage{
EPA.89b.loglead.df
}
\description{
Log-transformed lead concentrations (\eqn{\mu}g/L) at two background and four
  compliance wells (four monthly samples for each well).
}
\docType{data}
\format{
  A data frame with 24 observations on the following 4 variables.
  \describe{
    \item{\code{LogLead}}{Natural logarithm of lead concentrations (\eqn{\mu}g/L)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Well.type}}{a factor indicating the well type (background vs. compliance)}
  }
}
\source{
  USEPA. (1989b). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities, Interim Final Guidance}. EPA/530-SW-89-026. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C. p.5-11.
}
\keyword{datasets}
