% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.89b.toc.vec}
\alias{EPA.89b.toc.vec}
\title{Total Organic Carbon Concentrations from 1989 USEPA Guidance Document}
\usage{
EPA.89b.toc.vec
}
\description{
Numeric vector containing total organic carbon (TOC) concentrations (mg/L).
}
\docType{data}
\format{
  A numeric vector with 19 elements containing TOC concentrations (mg/L).
}
\source{
  USEPA. (1989b).
  \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities, Interim Final Guidance}.
  EPA/530-SW-89-026. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C. p.8-13.
}
\keyword{datasets}
