% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.92c.arsenic1.df}
\alias{EPA.92c.arsenic1.df}
\title{Arsenic Concentrations from 1992 USEPA Guidance Document}
\usage{
EPA.92c.arsenic1.df
}
\description{
Arsenic concentrations (ppm) at six monitoring wells (four monthly samples for each well).
}
\docType{data}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{Arsenic}}{Arsenic concentrations (ppm)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.  p.21.
}
\keyword{datasets}
