% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.92c.lognickel1.df}
\alias{EPA.92c.lognickel1.df}
\title{Log-transformed Nickel Concentrations from 1992 USEPA Guidance Document}
\usage{
EPA.92c.lognickel1.df
}
\description{
Log-transformed nickel concentrations (ppb) at four monitoring wells (five monthly samples for each well).
}
\docType{data}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{LogNickel}}{a numeric vector of log-transformed nickel concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.  p.15.
}
\keyword{datasets}
