% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.92c.nickel2.df}
\alias{EPA.92c.nickel2.df}
\title{Nickel Concentrations from 1992 USEPA Guidance Document}
\usage{
EPA.92c.nickel2.df
}
\description{
Nickel concentrations (ppb) at a monitoring well (eight months of samples, two samples for each sampling occasion).
}
\docType{data}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{Nickel}}{a numeric vector of nickel concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Sample}}{a factor indicating the sample (replicate) number}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.  p.78.
}
\keyword{datasets}
