% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{FcnsByCat}
\alias{FcnsByCat}
\title{EnvStats Functions Listed by Category}
\description{
Hyperlink list of \pkg{EnvStats} functions by category.
}
\alias{Functions By Category}
\concept{EnvStats Functions By Category}
\details{
  \itemize{
    \item \link[=FcnsByCatCalibration]{Calibration}
    \item \link[=FcnsByCatCensoredData]{Censored Data}
    \item \link[=FcnsByCatDataTrans]{Data Transformations}
    \item \link[=FcnsByCatEstDistParams]{Estimating Distribution Parameters}
    \item \link[=FcnsByCatEstDistQuants]{Estimating Distribution Quantiles}
    \item \link[=FcnsByCatGOFTests]{Goodness-of-Fit Tests}
    \item \link[=FcnsByCatHypothTests]{Hypothesis Tests}
    \item \link[=FcnsByCatMCandRisk]{Monte Carlo Simulation and Risk Assessment}
    \item \link[=FcnsByCatPlotProbDists]{Plotting Probability Distributions}
    \item \link[=FcnsByCatPlotUsingggplot2]{Plotting Using ggplot2}
    \item \link[=FcnsByCatPower]{Power and Sample Size Calculations}
    \item \link[=FcnsByCatPredInts]{Prediction Intervals}
    \item \link[=FcnsByCatPrintPlot]{Printing and Plotting Methods}
    \item \link[=FcnsByCatProbDists]{Probability Distributions and Random Numbers}
    \item \link[=FcnsByCatSumStats]{Summary Statistics}
    \item \link[=FcnsByCatTolInts]{Tolerance Intervals}
    \item \link[=FcnsByCatTrend]{Trend Analysis}
  }
}
\keyword{ package }
