% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{FcnsByCatCensoredData}
\alias{FcnsByCatCensoredData}
\title{EnvStats Functions for Censored Data}
\description{
The \pkg{EnvStats} functions listed below are useful for dealing with Type I censored data.
}
\details{
This help page has been shortened to keep function help focused on usage,
arguments, return values, and examples. Extended method details, formulas,
and background material are available in \code{vignette("extended-function-details", package = "EnvStats")}, section \code{FcnsByCatCensoredData}.
}
\alias{Censored Data}
\alias{EnvStats Functions for Censored Data}
\alias{Functions for Censored Data}
\concept{Less Than Detection Limit}
\keyword{ package }
