% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{FcnsByCatPlotProbDists}
\alias{FcnsByCatPlotProbDists}
\title{EnvStats Functions for Plotting Probability Distributions}
\description{
The \pkg{EnvStats} functions listed below are useful for plotting probability distributions.
}
\alias{Plotting Probability Distributions}
\alias{Plot Probability Distributions}
\details{
  \tabular{ll}{
  \emph{Function Name}                    \tab \emph{Description} \cr
  \code{\link{cdfCompare}}                \tab Plot two cumulative distribution functions with the same \eqn{x}-axis \cr
                                          \tab in order to compare them. \cr
  \code{\link{cdfPlot}}                   \tab Plot a cumulative distribution function. \cr
  \code{\link{ecdfPlot}}                  \tab Plot empirical cumulative distribution function. \cr
  \code{\link{epdfPlot}}                  \tab Plot empirical probability density function. \cr
  \code{\link{pdfPlot}}                   \tab Plot probability density function. \cr
  \code{\link{qqPlot}}                    \tab Produce a quantile-quantile (Q-Q) plot, also called a probability plot. \cr
  \code{\link{qqPlotGestalt}}             \tab Plot several Q-Q plots from the same distribution in order to \cr
                                          \tab develop a Gestalt of Q-Q plots for that distribution. \cr
 }
}
\keyword{ package }
