% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{FcnsByCatPlotUsingggplot2}
\alias{FcnsByCatPlotUsingggplot2}
\title{EnvStats Functions for Creating Plots Using the ggplot2 Package}
\description{
The \pkg{EnvStats} functions listed below are useful for creating plots with the
  \pkg{\link[ggplot2]{ggplot2}} package.
}
\alias{Plotting Using ggplot2}
\alias{Plot Using ggplot2}
\details{
  \tabular{ll}{
  \emph{Function Name}                    \tab \emph{Description} \cr
  \code{\link{geom_stripchart}}           \tab Adaptation of the \pkg{EnvStats} function \code{\link{stripChart}}, \cr
                                          \tab used to create a strip plot using functions from the package \cr
                                          \tab \pkg{\link[ggplot2]{ggplot2}}. \cr
  \code{\link{stat_n_text}}               \tab Add text indicating the sample size \cr
                                          \tab to a \pkg{\link[ggplot2]{ggplot2}} plot. \cr
  \code{\link{stat_mean_sd_text}}         \tab Add text indicating the mean and standard deviation \cr
                                          \tab to a \pkg{\link[ggplot2]{ggplot2}} plot. \cr
  \code{\link{stat_median_iqr_text}}      \tab Add text indicating the median and interquartile range \cr
                                          \tab to a \pkg{\link[ggplot2]{ggplot2}} plot. \cr
  \code{\link{stat_test_text}}            \tab Add text indicating the results of a hypothesis test \cr
                                          \tab comparing groups to a \pkg{\link[ggplot2]{ggplot2}} plot. \cr
 }
}
\keyword{ package }
