% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predIntNparSimultaneousTestPower.R
\name{predIntNparSimultaneousTestPower}
\alias{predIntNparSimultaneousTestPower}
\title{Probability That at Least One Set of Future Observations Violates the Given Rule Based on a Nonparametric Simultaneous Prediction Interval}
\usage{
predIntNparSimultaneousTestPower(n, n.median = 1, k = 1, m = 2, r = 1,
    rule = "k.of.m", lpl.rank = ifelse(pi.type == "upper", 0, 1),
    n.plus.one.minus.upl.rank = ifelse(pi.type == "lower", 0, 1),
    delta.over.sigma = 0, pi.type = "upper", r.shifted = r,
    method = "approx", NMC = 100, ci = FALSE, ci.conf.level = 0.95,
    integrate.args.list = NULL, evNormOrdStats.method = "royston")
}
\description{
Compute the probability that at least one set of future observations violates the
  given rule based on a nonparametric simultaneous prediction interval for the next
  \eqn{r} future sampling occasions.  The three possible rules are:
  \eqn{k}-of-\eqn{m}, California, or Modified California.  The probability is based
  on assuming the true distribution of the observations is \link[stats:Normal]{normal}.
}
\details{
This help page has been shortened to keep function help focused on usage,
arguments, return values, and examples. Extended method details, formulas,
and background material are available in \code{vignette("extended-function-details", package = "EnvStats")}, section \code{predIntNparSimultaneousTestPower}.
}
\arguments{
  \item{n}{
  vector of positive integers specifying the sample sizes.
  Missing (\code{NA}), undefined (\code{NaN}), and infinite (\code{Inf}, \code{-Inf})
  values are not allowed.
}
  \item{n.median}{
  vector of positive odd integers specifying the sample size associated with the
  future medians.  The default value is \code{n.median=1} (i.e., individual
  observations).  Note that all future medians must be based on the same
  sample size.
}
  \item{k}{
  for the \eqn{k}-of-\eqn{m} rule (\code{rule="k.of.m"}), a vector of positive integers
  specifying the minimum number of observations (or medians) out of \eqn{m}
  observations (or medians) (all obtained on one future sampling \dQuote{occassion})
  the prediction interval should contain.
  The default value is \code{k=1}.  This argument is ignored when the argument
  \code{rule} is not equal to \code{"k.of.m"}.
}
  \item{m}{
  vector of positive integers specifying the maximum number of future observations (or
  medians) on one future sampling \dQuote{occasion}.
  The default value is \code{m=2}, except when \code{rule="Modified.CA"}, in which
  case this argument is ignored and \code{m} is automatically set equal to \code{4}.
}
  \item{r}{
  vector of positive integers specifying the number of future sampling
  \dQuote{occasions}.  The default value is \code{r=1}.
}
  \item{rule}{
  character string specifying which rule to use.  The possible values are
  \code{"k.of.m"} (\eqn{k}-of-\eqn{m} rule; the default), \code{"CA"} (California rule),
  and \code{"Modified.CA"} (modified California rule).
}
  \item{lpl.rank}{
  vector of non-negative integers indicating the rank of the order statistic to use for
  the lower bound of the prediction interval.  When \code{pi.type="lower"}, the
  default value is \code{lpl.rank=1} (implying the minimum value of \code{x} is used
  as the lower bound of the prediction interval).  When \code{pi.type="upper"},
  the argument \code{lpl.rank} is set equal to \code{0}.
}
  \item{n.plus.one.minus.upl.rank}{
  vector of non-negative integers related to the rank of the order statistic to use for
  the upper bound of the prediction interval.  A value of \code{n.plus.one.minus.upl.rank=1}
  (the default) means use the first largest value, and in general a value of \cr
  \code{n.plus.one.minus.upl.rank=}\eqn{i} means use the \eqn{i}'th largest value.
  When \code{pi.type="lower"}, the argument \code{n.plus.one.minus.upl.rank} is set
  equal to \code{0}.
}
  \item{delta.over.sigma}{
  numeric vector indicating the ratio \eqn{\Delta/\sigma}.  The quantity
  \eqn{\Delta} (delta) denotes the difference between the mean of the population
  that was sampled to construct the prediction interval, and the mean of the
  population that will be sampled to produce the future observations.  The quantity
  \eqn{\sigma} (sigma) denotes the population standard deviation for both populations.
  The default value is \cr
  \code{delta.over.sigma=0}.
}
  \item{pi.type}{
  character string indicating what kind of prediction interval to compute.
  The possible values are \code{"two.sided"} (the default), \code{"lower"}, and
  \code{"upper"}.
}
  \item{r.shifted}{
  vector of positive integers specifying the number of future sampling occasions for
  which the scaled mean is shifted by \eqn{\Delta/\sigma}.  All values must be
  integeters between \code{1} and the corresponding element of \code{r}.
  The default value is \code{r.shifted=r}.
}
  \item{method}{
  character string indicating what method to use to compute the power.  The possible
  values are \code{"approx"} (approximation based on \cr
  \code{\link{predIntNormSimultaneousTestPower}}; the default) and
  \code{"simulate"} (Monte Carlo simulation).
}
  \item{NMC}{
  positive integer indicating the number of Monte Carlo trials to run when \cr
  \code{method="simulate"}.  The default value is \code{NMC=100}.
}
  \item{ci}{
  logical scalar indicating whether to compute a confidence interval for the power
  when \code{method="simulate"}.  The default value is \code{ci=FALSE}.
}
  \item{ci.conf.level}{
  numeric scalar between 0 and 1 indicating the confidence level associated with the
  confidence interval for the power.  The argument is ignored if \code{ci=FALSE}
  or \code{method="approx"}.
}
  \item{integrate.args.list}{
  list of arguments to supply to the \code{\link{integrate}} function.  The default
  value is \code{NULL}.
}
  \item{evNormOrdStats.method}{
  character string indicating which method to use in the call to
  \code{\link{evNormOrdStatsScalar}} when \code{method="approx"}.  The default
  value is \code{evNormOrdStats.method="royston"}.  See the DETAILS section for
  more information.
}
}
\value{
  vector of values between 0 and 1 equal to the probability that
  the rule will be violated.
}
\references{
  See the help file for \code{\link{predIntNparSimultaneous}}.

  Gansecki, M. (2009).  \emph{Using the Optimal Rank Values Calculator}.
  US Environmental Protection Agency, Region 8, March 10, 2009.
}
\author{
    Steven P. Millard (\email{EnvStats@ProbStatInfo.com})
}
\note{
  See the help file for \code{\link{predIntNparSimultaneous}}.

  In the course of designing a sampling program, an environmental scientist may wish
  to determine the relationship between sample size, significance level, power, and
  scaled difference if one of the objectives of the sampling program is to determine
  whether two distributions differ from each other.  The functions
  \code{predIntNparSimultaneousTestPower} and \cr
  \code{\link{plotPredIntNparSimultaneousTestPowerCurve}} can be
  used to investigate these relationships for the case of normally-distributed
  observations.
}
\seealso{
  \code{\link{plotPredIntNparSimultaneousTestPowerCurve}},
  \code{\link{predIntNparSimultaneous}}, \cr
  \code{\link{predIntNparSimultaneousN}},
  \code{\link{predIntNparSimultaneousConfLevel}}, \cr
  \code{\link{plotPredIntNparSimultaneousDesign}},
  \code{\link{predIntNpar}}, \code{\link{tolIntNpar}}.
}
\examples{
  # Example 19-5 of USEPA (2009, p. 19-33) shows how to compute nonparametric upper
  # simultaneous prediction limits for various rules based on trace mercury data (ppb)
  # collected in the past year from a site with four background wells and 10 compliance
  # wells (data for two of the compliance wells  are shown in the guidance document).
  # The facility must monitor the 10 compliance wells for five constituents
  # (including mercury) annually.

  # Here we will compute the confidence levels and powers associated with
  # two different sampling plans:
  # 1) the 1-of-2 retesting plan for a median of order 3 using the
  #    background maximum and
  # 2) the 1-of-4 plan on individual observations using the 3rd highest
  #    background value.
  # Power will be computed assuming a normal distribution and setting
  # delta.over.sigma equal to 2, 3, and 4.
  # The data for this example are stored in EPA.09.Ex.19.5.mercury.df.

  # We will pool data from 4 background wells that were sampled on
  # a number of different occasions, giving us a sample size of
  # n = 20 to use to construct the prediction limit.

  # There are 10 compliance wells and we will monitor 5 different
  # constituents at each well annually.  For this example, USEPA (2009)
  # recommends setting r to the product of the number of compliance wells and
  # the number of evaluations per year.

  # To determine the minimum confidence level we require for
  # the simultaneous prediction interval, USEPA (2009) recommends
  # setting the maximum allowed individual Type I Error level per constituent to:

  # 1 - (1 - SWFPR)^(1 / Number of Constituents)

  # which translates to setting the confidence limit to

  # (1 - SWFPR)^(1 / Number of Constituents)

  # where SWFPR = site-wide false positive rate.  For this example, we
  # will set SWFPR = 0.1.  Thus, the required individual Type I Error level
  # and confidence level per constituent are given as follows:

  # n  = 20 based on 4 Background Wells
  # nw = 10 Compliance Wells
  # nc =  5 Constituents
  # ne =  1 Evaluation per year

  n  <- 20
  nw <- 10
  nc <-  5
  ne <-  1

  # Set number of future sampling occasions r to
  # Number Compliance Wells x Number Evaluations per Year
  r  <-  nw * ne

  conf.level <- (1 - 0.1)^(1 / nc)
  conf.level
  #[1] 0.9791484

  # So the required confidence level is 0.98, or 98%.
  # Now determine the confidence level associated with each plan.
  # Note that both plans achieve the required confidence level.

  # 1) the 1-of-2 retesting plan for a median of order 3 using the
  #    background maximum

  predIntNparSimultaneousConfLevel(n = 20, n.median = 3, k = 1, m = 2, r = r)
  #[1] 0.9940354


  # 2) the 1-of-4 plan based on individual observations using the 3rd highest
  #    background value.

  predIntNparSimultaneousConfLevel(n = 20, k = 1, m = 4, r = r,
    n.plus.one.minus.upl.rank = 3)
  #[1] 0.9864909

  #------------------------------------------------------------------------------
  # Compute approximate power of each plan to detect contamination at just 1 well
  # assuming true underying distribution of Hg is Normal at all wells and
  # using delta.over.sigma equal to 2, 3, and 4.
  #------------------------------------------------------------------------------

  # Computer aproximate power for
  # 1) the 1-of-2 retesting plan for a median of order 3 using the
  #    background maximum

  predIntNparSimultaneousTestPower(n = 20, n.median = 3, k = 1, m = 2, r = r,
    delta.over.sigma = 2:4, r.shifted = 1)
  #[1] 0.3953712 0.9129671 0.9983054


  # Compute approximate power for
  # 2) the 1-of-4 plan based on individual observations using the 3rd highest
  #    background value.

  predIntNparSimultaneousTestPower(n = 20, k = 1, m = 4, r = r,
    n.plus.one.minus.upl.rank = 3, delta.over.sigma = 2:4, r.shifted = 1)
  #[1] 0.4367972 0.8694664 0.9888779


  #----------

  \dontrun{
  # Compare estimated power using approximation method with estimated power
  # using Monte Carlo simulation for the 1-of-4 plan based on individual
  # observations using the 3rd highest background value.

  predIntNparSimultaneousTestPower(n = 20, k = 1, m = 4, r = r,
    n.plus.one.minus.upl.rank = 3, delta.over.sigma = 2:4, r.shifted = 1,
    method = "simulate", ci = TRUE, NMC = 1000)
  #[1] 0.437 0.863 0.989
  #attr(,"conf.int")
  #         [,1]      [,2]      [,3]
  #LCL 0.4111999 0.8451148 0.9835747
  #UCL 0.4628001 0.8808852 0.9944253
  }

  #==========

  # Cleanup
  #--------
  rm(n, nw, nc, ne, r, conf.level)
}
\keyword{ distribution }
\keyword{ htest }
