% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{print.htestEnvStats}
\alias{print.htestEnvStats}
\title{Print Output of EnvStats Hypothesis Tests}
\usage{
\method{print}{htestEnvStats}(x, ...)
}
\description{
Print objects of class \code{"htestEnvStats"} by a simple
  \code{\link[base:print]{print}} method.

  The function \code{print.htestEnvStats} formats and prints the results of
  performing a hypothesis test using one of the functions listed in the help
  file for \code{\link{htestEnvStats.object}}. This method is automatically
  called by the standard \R \code{\link[base:print]{print}} generic when given
  an object of class \code{"\link[=htestEnvStats.object]{htestEnvStats}"}.
}
\arguments{
  \item{x}{
  an object of class \code{"htestEnvStats"}. See
  \code{\link{htestEnvStats.object}} for details.
}
  \item{\dots}{
  arguments that can be supplied to the \code{\link[base]{format}} function.
}
}
\details{
  This method prints null and alternative hypotheses, name of the test,
  estimated population parameter(s) involved in the null hypothesis, estimation
  method (if present), data name, sample size (if present), number of missing
  observations removed prior to performing the test (if present), value of the
  test statistic, parameters associated with the null distribution of the test
  statistic, p-value associated with the test statistic, and confidence
  interval for the population parameter (if present).
}
\value{
  Invisibly returns the input \code{x}.
}
\references{
  Chambers, J. M. and Hastie, T. J. (1992).  \emph{Statistical Models in S}.
  Wadsworth & Brooks/Cole.
}
\author{
  Steven P. Millard (\email{EnvStats@ProbStatInfo.com})
}
\seealso{
  \link{Hypothesis Tests}, \code{\link{htestEnvStats.object}},
  \code{\link[base:print]{print}}.
}
\examples{
  # Create an object of class "htestEnvStats", then print it out.
  #--------------------------------------------------------------
  htestEnvStats.obj <- chenTTest(EPA.02d.Ex.9.mg.per.L.vec, mu = 30)

  mode(htestEnvStats.obj)
  #[1] "list"

  class(htestEnvStats.obj)
  #[1] "htestEnvStats"

  names(htestEnvStats.obj)
  # [1] "statistic"   "parameters"  "p.value"     "estimate"
  # [5] "null.value"  "alternative" "method"      "sample.size"
  # [9] "data.name"   "bad.obs"     "interval"

  htestEnvStats.obj

  #==========

  # Extract the test statistic
  #---------------------------

  htestEnvStats.obj$statistic

  #==========

  # Clean up
  #---------
  rm(htestEnvStats.obj)
}
\keyword{print}
