% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialCorrelationTest.R
\name{serialCorrelationTest}
\alias{serialCorrelationTest}
\alias{serialCorrelationTest.default}
\alias{serialCorrelationTest.lm}
\title{Test for the Presence of Serial Correlation}
\usage{
serialCorrelationTest(x, ...)

  \method{serialCorrelationTest}{default}(x, test = "rank.von.Neumann",
    alternative = "two.sided", conf.level = 0.95, ...)

  \method{serialCorrelationTest}{lm}(x, test = "rank.von.Neumann",
    alternative = "two.sided", conf.level = 0.95, ...)
}
\description{
\code{serialCorrelationTest} is a generic function used to test for the
  presence of lag-one serial correlation using either the rank
  von Neumann ratio test, the normal approximation based on the Yule-Walker
  estimate of lag-one correlation, or the normal approximation based on the
  MLE of lag-one correlation.  The function invokes particular
  \code{\link{methods}} which depend on the \code{\link{class}} of the first
  argument.

  Currently, there is a default method and a method for objects of class \code{"lm"}.
}
\details{
This help page has been shortened to keep function help focused on usage,
arguments, return values, and examples. Extended method details, formulas,
and background material are available in \code{vignette("extended-function-details", package = "EnvStats")}, section \code{serialCorrelationTest}.
}
\alias{serial correlation test}
\alias{serial correlation}
\arguments{
  \item{x}{
  numeric vector of observations, a numeric univariate time series of
  class \code{"ts"}, or an object of class \code{"lm"}.  Undefined (\code{NaN}) and
  infinite (\code{Inf}, \code{-Inf}) values are not allowed for \code{x}
  when \code{x} is a numeric vector or time series, nor for the residuals
  associated with \code{x} when \code{x} is an object of class \code{"lm"}.

  When \code{test="AR1.mle"}, missing (\code{NA}) values are allowed, otherwise
  they are not allowed.  When \code{x} is a numeric vector of observations
  or a numeric univariate time series of class \code{"ts"}, it must contain at least
  3 non-missing values.  When \code{x} is an object of class \code{"lm"}, the
  residuals must contain at least 3 non-missing values.

  Note:  when \code{x} is an object of class \code{"lm"}, the linear model
  should have been fit using the argument \code{na.action=na.exclude} in the
  call to \code{\link{lm}} in order to correctly deal with missing values.
}
  \item{test}{
  character string indicating which test to use.  The possible values are: \cr
  \code{"rank.von.Neumann"} (rank von Neumann ratio test; the default), \cr
  \code{"AR1.yw"} (z-test based on Yule-Walker lag-one estimate of correlation), and \cr
  \code{"AR1.mle"} (z-test based on MLE of lag-one correlation).
}
  \item{alternative}{
  character string indicating the kind of alternative hypothesis.  The possible
  values are \code{"two.sided"} (the default), \code{"greater"}, and \code{"less"}.
}
  \item{conf.level}{
  numeric scalar between 0 and 1 indicating the confidence level associated with
  the confidence interval for the population lag-one autocorrelation.  The default
  value is \code{conf.level=0.95}.
}
  \item{\dots}{
  optional arguments for possible future methods.  Currently not used.
}
}
\value{
  A list of class \code{"htestEnvStats"} containing the results of the hypothesis test.
  See the help file for \code{\link{htestEnvStats.object}} for details.
}
\references{
  Bartels, R. (1982).  The Rank Version of von Neumann's Ratio Test for Randomness.
  \emph{Journal of the American Statistical Association} \bold{77}(377), 40--46.

  Berthouex, P.M., and L.C. Brown. (2002).
  \emph{Statistics for Environmental Engineers}.  Second Edition.
  Lewis Publishers, Boca Raton, FL.

  Box, G.E.P., and G.M. Jenkins. (1976).
  \emph{Time Series Analysis: Forecasting and Control}.  Prentice Hall,
  Englewood Cliffs, NJ, Chapter 2.

  Cox, D.R. (1966).  The Null Distribution of the First Serial Correlation Coefficient.
  \emph{Biometrika} \bold{53}, 623--626.

  Draper, N., and H. Smith. (1998).  \emph{Applied Regression Analysis}.
  Third Edition. John Wiley and Sons, New York, pp.69-70;181-192.

  Durbin, J., and G.S. Watson. (1950).  Testing for Serial Correlation in Least
  Squares Regression I.  \emph{Biometrika} \bold{37}, 409--428.

  Durbin, J., and G.S. Watson. (1951).  Testing for Serial Correlation in Least
  Squares Regression II.  \emph{Biometrika} \bold{38}, 159--178.

  Durbin, J., and G.S. Watson. (1971). Testing for Serial Correlation in Least Squares
  Regression III.  \emph{Biometrika} \bold{58}, 1--19.

  Helsel, D.R., and R.M. Hirsch. (1992).  \emph{Statistical Methods in Water
  Resources Research}.  Elsevier, New York, NY, pp.250--253.

  Johnson, N. L., S. Kotz, and N. Balakrishnan. (1995).  \emph{Continuous Univariate
  Distributions, Volume 2}.  Second Edition.  John Wiley and Sons, New York,
  Chapter 25.

  Knoke, J.D. (1975).  Testing for Randomness Against Autocorrelation Alternatives:
  The Parametric Case.  \emph{Biometrika} \bold{62}, 571--575.

  Knoke, J.D. (1977).  Testing for Randomness Against Autocorrelation Alternatives:
  Alternative Tests.  \emph{Biometrika} \bold{64}, 523--529.

  Lehmann, E.L. (1975).  \emph{Nonparametrics:  Statistical Methods Based on Ranks}.
  Holden-Day, Oakland, CA, 457pp.

  von Neumann, J., R.H. Kent, H.R. Bellinson, and B.I. Hart. (1941).  The Mean Square
  Successive Difference.  \emph{Annals of Mathematical Statistics} \bold{12}(2),
  153--162.

  Wald, A., and J. Wolfowitz. (1943).  An Exact Test for Randomness in the
  Non-Parametric Case Based on Serial Correlation.  \emph{Annals of Mathematical
  Statistics} \bold{14}, 378--388.
}
\author{
  Steven P. Millard (\email{EnvStats@ProbStatInfo.com})
}
\note{
  Data collected over time on the same phenomenon are called a time series.
  A time series is usually modeled as a single realization of a stochastic process;
  that is, if we could go back in time and repeat the experiment, we would get
  different results that would vary according to some probabilistic law.
  The simplest kind of time series is a stationary time series, in which the mean
  value is constant over time, the variability of the observations is constant over
  time, etc.  That is, the probability distribution associated with each future
  observation is the same.

  A common concern in applying standard statistical tests to time series data is
  the assumption of independence.  Most conventional statistical hypothesis tests
  assume the observations are independent, but data collected sequentially in time
  may not satisfy this assumption.  For example, high observations may tend to
  follow high observations (positive serial correlation), or low observations may
  tend to follow high observations (negative serial correlation).  One way to
  investigate the assumption of independence is to estimate the lag-one serial
  correlation and test whether it is significantly different from 0.

  The null distribution of the serial correlation coefficient may be badly affected
  by departures from normality in the underlying process (Cox, 1966; Bartels, 1977).
  It is therefore a good idea to consider using a nonparametric test for randomness
  if the normality of the underlying process is in doubt (Bartels, 1982).
  Knoke (1977) showed that under normality, the test based on the rank serial
  correlation coefficient (and hence the test based on the rank von Neumann ratio
  statistic) has asymptotic relative efficiency of 0.91 with respect to using the
  test based on the ordinary serial correlation coefficient against the alternative
  of first-order autocorrelation.

  Bartels (1982) performed an extensive simulation study of the power of the
  rank von Neumann ratio test relative to the standard von Neumann ratio test
  (based on the statistic in Equation (9) above) and the runs test
  (Lehmann, 1975, 313-315).  He generated a first-order autoregressive process for
  sample sizes of 10, 25, and 50, using 6 different parent distributions: normal,
  Cauchy, contaminated normal, Johnson, Stable, and exponential.  Values of
  lag-1 autocorrelation ranged from -0.8 to 0.8.  Bartels (1982) found three
  important results:
  \itemize{
  \item The rank von Neumann ratio test is far more powerful than the runs test.

  \item For the normal process, the power of the rank von Neumann ratio test was
  never less than 89\% of the power of the standard von Neumann ratio test.

  \item For non-normal processes, the rank von Neumann ratio test was often much
  more powerful than of the standard von Neumann ratio test.
}
}
\seealso{
  \code{\link{htestEnvStats.object}}, \code{\link{acf}}, \code{\link{ar}},
  \code{\link{arima}}, \code{\link{arima.sim}},
  \code{\link{ts.plot}}, \code{\link{plot.ts}},
  \code{\link{lag.plot}}, \link{Hypothesis Tests}.
}
\examples{
  # Generate a purely random normal process, then use serialCorrelationTest
  # to test for the presence of correlation.
  # (Note: the call to set.seed allows you to reproduce this example.)

  set.seed(345)
  x <- rnorm(100)

  # Look at the data
  #-----------------
  dev.new()
  ts.plot(x)

  dev.new()
  acf(x)

  # Test for serial correlation
  #----------------------------
  serialCorrelationTest(x)

  #Results of Hypothesis Test
  #--------------------------
  #
  #Null Hypothesis:                 rho = 0
  #
  #Alternative Hypothesis:          True rho is not equal to 0
  #
  #Test Name:                       Rank von Neumann Test for
  #                                 Lag-1 Autocorrelation
  #                                 (Beta Approximation)
  #
  #Estimated Parameter(s):          rho = 0.02773737
  #
  #Estimation Method:               Yule-Walker
  #
  #Data:                            x
  #
  #Sample Size:                     100
  #
  #Test Statistic:                  RVN = 1.929733
  #
  #P-value:                         0.7253405
  #
  #Confidence Interval for:         rho
  #
  #Confidence Interval Method:      Normal Approximation
  #
  #Confidence Interval Type:        two-sided
  #
  #Confidence Level:                95%
  #
  #Confidence Interval:             LCL = -0.1681836
  #                                 UCL =  0.2236584

  # Clean up
  #---------
  rm(x)
  graphics.off()

  #==========

  # Now use the R function arima.sim to generate an AR(1) process with a
  # lag-1 autocorrelation of 0.8, then test for autocorrelation.

  set.seed(432)
  y <- arima.sim(model = list(ar = 0.8), n = 100)

  # Look at the data
  #-----------------
  dev.new()
  ts.plot(y)

  dev.new()
  acf(y)

  # Test for serial correlation
  #----------------------------
  serialCorrelationTest(y)

  #Results of Hypothesis Test
  #--------------------------
  #
  #Null Hypothesis:                 rho = 0
  #
  #Alternative Hypothesis:          True rho is not equal to 0
  #
  #Test Name:                       Rank von Neumann Test for
  #                                 Lag-1 Autocorrelation
  #                                 (Beta Approximation)
  #
  #Estimated Parameter(s):          rho = 0.835214
  #
  #Estimation Method:               Yule-Walker
  #
  #Data:                            y
  #
  #Sample Size:                     100
  #
  #Test Statistic:                  RVN = 0.3743174
  #
  #P-value:                         0
  #
  #Confidence Interval for:         rho
  #
  #Confidence Interval Method:      Normal Approximation
  #
  #Confidence Interval Type:        two-sided
  #
  #Confidence Level:                95%
  #
  #Confidence Interval:             LCL = 0.7274307
  #                                 UCL = 0.9429973

  #----------

  # Clean up
  #---------
  rm(y)
  graphics.off()

  #==========

  # The data frame Air.df contains information on ozone (ppb^1/3),
  # radiation (langleys), temperature (degrees F), and wind speed (mph)
  # for 153 consecutive days between May 1 and September 30, 1973.
  # First test for serial correlation in (the cube root of) ozone.
  # Note that we must use the test based on the MLE because the time series
  # contains missing values.  Serial correlation appears to be present.
  # Next fit a linear model that includes the predictor variables temperature,
  # radiation, and wind speed, and test for the presence of serial correlation
  # in the residuals.  There is no evidence of serial correlation.

  # Look at the data
  #-----------------

  Air.df
  #              ozone radiation temperature wind
  #05/01/1973 3.448217       190          67  7.4
  #05/02/1973 3.301927       118          72  8.0
  #05/03/1973 2.289428       149          74 12.6
  #05/04/1973 2.620741       313          62 11.5
  #05/05/1973       NA        NA          56 14.3
  #...
  #09/27/1973       NA       145          77 13.2
  #09/28/1973 2.410142       191          75 14.3
  #09/29/1973 2.620741       131          76  8.0
  #09/30/1973 2.714418       223          68 11.5

  #----------

  # Test for serial correlation
  #----------------------------

  with(Air.df,
    serialCorrelationTest(ozone, test = "AR1.mle"))

  #Results of Hypothesis Test
  #--------------------------
  #
  #Null Hypothesis:                 rho = 0
  #
  #Alternative Hypothesis:          True rho is not equal to 0
  #
  #Test Name:                       z-Test for
  #                                 Lag-1 Autocorrelation
  #                                 (Wald Test Based on MLE)
  #
  #Estimated Parameter(s):          rho = 0.5641616
  #
  #Estimation Method:               Maximum Likelihood
  #
  #Data:                            ozone
  #
  #Sample Size:                     153
  #
  #Number NA/NaN/Inf's:             37
  #
  #Test Statistic:                  z = 7.586952
  #
  #P-value:                         3.28626e-14
  #
  #Confidence Interval for:         rho
  #
  #Confidence Interval Method:      Normal Approximation
  #
  #Confidence Interval Type:        two-sided
  #
  #Confidence Level:                95%
  #
  #Confidence Interval:             LCL = 0.4184197
  #                                 UCL = 0.7099034

  #----------

  # Next fit a linear model that includes the predictor variables temperature,
  # radiation, and wind speed, and test for the presence of serial correlation
  # in the residuals.  Note setting the argument na.action = na.exclude in the
  # call to lm to correctly deal with missing values.
  #----------------------------------------------------------------------------

  lm.ozone <- lm(ozone ~ radiation + temperature + wind +
    I(temperature^2) + I(wind^2),
    data = Air.df, na.action = na.exclude)


  # Now test for serial correlation in the residuals.
  #--------------------------------------------------

  serialCorrelationTest(lm.ozone, test = "AR1.mle")

  #Results of Hypothesis Test
  #--------------------------
  #
  #Null Hypothesis:                 rho = 0
  #
  #Alternative Hypothesis:          True rho is not equal to 0
  #
  #Test Name:                       z-Test for
  #                                 Lag-1 Autocorrelation
  #                                 (Wald Test Based on MLE)
  #
  #Estimated Parameter(s):          rho = 0.1298024
  #
  #Estimation Method:               Maximum Likelihood
  #
  #Data:                            Residuals
  #
  #Data Source:                     lm.ozone
  #
  #Sample Size:                     153
  #
  #Number NA/NaN/Inf's:             42
  #
  #Test Statistic:                  z = 1.285963
  #
  #P-value:                         0.1984559
  #
  #Confidence Interval for:         rho
  #
  #Confidence Interval Method:      Normal Approximation
  #
  #Confidence Interval Type:        two-sided
  #
  #Confidence Level:                95%
  #
  #Confidence Interval:             LCL = -0.06803223
  #                                 UCL =  0.32763704

  # Clean up
  #---------
  rm(lm.ozone)

}
\keyword{htestEnvStats}
\keyword{univar}
\keyword{models}
