% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{Total.P.df}
\alias{Total.P.df}
\title{Total Phosphorus Data from Chesapeake Bay}
\usage{
Total.P.df
}
\description{
Monthly estimated total phosphorus mass (mg) within a water column at two different
  stations for the 5-year time period October 1984 to September 1989 from a study on
  phosphorus concentration conducted in the Chesapeake Bay.
}
\docType{data}
\format{
  A data frame with 60 observations on the following 4 variables.
  \describe{
    \item{\code{CB3.1}}{a numeric vector of phosphorus concentrations at station CB3.1}
    \item{\code{CB3.3e}}{a numeric vector phosphorus concentrations at station CB3.3e}
    \item{\code{Month}}{a factor indicating the month the observation was taken}
    \item{\code{Year}}{a numeric vector indicating the year an observation was taken}
  }
}
\source{
  Neerchal, N. K., and S. L. Brunenmeister. (1993). Estimation of Trend in Chesapeake Bay Water Quality Data.
  In Patil, G.P., and C.R. Rao, eds., \emph{Handbook of Statistics, Vol. 6: Multivariate Environmental Statistics}.
  North-Holland, Amsterdam, Chapter 19, 407-422.
}
\keyword{datasets}
