% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{Distribution.df}
\alias{Distribution.df}
\title{Data Frame Summarizing Available Probability Distributions and Estimation Methods}
\usage{
Distribution.df
}
\description{
Data frame summarizing information about available probability
  distributions in \R and the \pkg{EnvStats} package, and which
  distributions have associated functions for estimating distribution
  parameters.
}
\details{
This help page has been shortened to keep function help focused on usage,
arguments, return values, and examples. Extended method details, formulas,
and background material are available in \code{vignette("extended-function-details", package = "EnvStats")}, section \code{Distribution.df}.
}
\docType{data}
\format{
  A data frame with 35 rows corresponding to 35 different available
  probability distributions, and 25 columns containing information
  associated with these probability distributions.
  \describe{
    \item{\code{Name}}{a character vector containing the name of
      the probability distribution (see the column labeled \bold{Name}
      in the table below).}
    \item{\code{Type}}{a character vector indicating the type of
      distribution (see the column labeled \bold{Type} in the table below).
      Possible values are \code{"Finite Discrete"}, \code{"Discrete"},
      \code{"Continuous"}, and \code{"Mixed"}.}
    \item{\code{Support.Min}}{a character vector indicating the minimum value
      the random variable can assume (see the column labeled \bold{Range} in
      the table below).  The reason this is a character vector instead of a
      numeric vector is because some distributions have a lower bound that
      depends on the value of a distribution parameter.  For example,
      the minimum value for a \link{Uniform} distribution is given by the
      value of the parameter \code{min}.}
    \item{\code{Support.Max}}{a character vector indicating the maximum value
      the random variable can assume (see the column labeled \bold{Range} in
      the table below).  The reason this is a character vector instead of a
      numeric vector is because some distributions have an upper bound that
      depends on the value of a distribution parameter.  For example,
      the maximum value for a \link{Uniform} distribution is given by the value
      of the parameter \code{max}.}
    \item{\code{Estimation.Method(s)}}{a character vector indicating the
      names of the methods available to estimate the distribution parameter(s)
      (see the column labeled \bold{Estimation Method(s)} in the table below).
      Possible values include \code{"mle"} (maximum likelihood), \code{"mme"}
      (method of moments), \code{"mmue"} (method of moments based on the
      unbiased estimate of variance), \code{"mvue"} (minimum variance unbiased),
      \code{"qmle"} (quasi-mle), etc., or some combination of these.  In
      cases where an estimator is more than one kind, a slash (\code{/}) is
      used to denote all methods covered by the single estimator.  For example,
      for the Binomial distribution, the sample proportion is the maximum
      likelihood, method of moments, and minimum variance unbiased estimator,
      so this method is denoted as \code{"mle/mme/mvue"}.  See the help files
      for the specific function listed under
      \link[=FcnsByCatEstDistParams]{Estimating Distribution Parameters} for an
      explanation of each of these estimation methods.}
    \item{\code{Quantile.Estimation.Method(s)}}{a character vector indicating
      the names of the methods available to estimate the distribution
      quantiles.  For many distributions, these are the same as
      \code{Estimation.Method(s)}.  See the help files for the specific
      function listed under
      \link[=FcnsByCatEstDistQuants]{Estimating Distribution Quantiles} for an
      explanation of each of these estimation methods.}
    \item{\code{Prediction.Interval.Method(s)}}{a character vector indicating
      the names of the methods available to create prediction intervals.  See
      the help files for the specific function listed under
      \link[=FcnsByCatPredInts]{Prediction Intervals} for an explanation of each of
      these estimation methods.}
    \item{\code{Singly.Censored.Estimation.Method(s)}}{a character vector
      indicating the names of the methods available to estimate the
      distribution parameter(s) for Type I singly-censored data.  See the
      help files for the specific function listed under
      \emph{Estimating Distribution Parameters} in the help file for
      \link[=FcnsByCatCensoredData]{Censored Data}
      for an explanation of each of these estimation methods.}
    \item{\code{Multiply.Censored.Estimation.Method(s)}}{a character vector
      indicating the names of the methods available to estimate the
      distribution parameter(s) for Type I multiply-censored data.  See the
      help files for the specific function listed under
      \emph{Estimating Distribution Parameters} in the help file for
      \link[=FcnsByCatCensoredData]{Censored Data}
      for an explanation of each of these estimation methods.}
    \item{\code{Number.parameters}}{a numeric vector indicating the number of
      parameters associated with the distribution (see the column labeled
      \bold{Parameters} in the table below).}
    \item{\code{Parameter.1}}{the columns labeled
      \code{Parameter.1}, \code{Parameter.2}, \ldots, \code{Parameter.5} are
      character vectors containing the names of the distribution parameters
      (see the column labeled \bold{Parameters} in the table below).  If a
      distribution has \eqn{n} parameters and \eqn{n < 5}, then the columns
      labeled \code{Parameter.n+1}, \ldots, \code{Parameter.5} are empty.  For
      example, the \link{Normal} distribution has only two parameters
      associated with it (\code{mean} and \code{sd}), so the fields in
      \code{Parameter.3}, \code{Parameter.4}, and \code{Parameter.5} are
      empty.}
    \item{\code{Parameter.2}}{see \code{Parameter.1}}
    \item{\code{Parameter.3}}{see \code{Parameter.1}}
    \item{\code{Parameter.4}}{see \code{Parameter.1}}
    \item{\code{Parameter.5}}{see \code{Parameter.1}}
    \item{\code{Parameter.1.Min}}{the columns labeled \code{Parameter.1.Min},
      \code{Parameter.2.Min}, \ldots, \cr
      \code{Parameter.5.Min} are character
      vectors containing the minimum values that can be assumed by the
      distribution parameters (see the column labeled \bold{Parameter Range(s)}
      in the table below).

      The reason these are character vectors instead of numeric vectors is
      because some parameters have a lower bound of \code{0} but must be
      strictly bigger than \code{0} (e.g., the parameter \code{sd} for the
      \link{Normal} distribution), in which case the lower bound is
      \code{.Machine$double.eps}, which may vary from machine to machine.
      Also, some parameters have a lower bound that depends on the value of
      another parameter.  For example, the parameter \code{max} for a
      \link{Uniform} distribution is bounded below by the value of the
      parameter \code{min}.

      If a distribution has \eqn{n} parameters and \eqn{n < 5}, then the
      columns labeled \code{Parameter.n+1.Min}, \ldots, \code{Parameter.5.Min}
      have the missing value code (\code{NA}).  For example, the \link{Normal}
      distribution has only two parameters associated with it (\code{mean}
      and \code{sd}) so the fields in \cr
      \code{Parameter.3.Min}, \code{Parameter.4.Min}, and \code{Parameter.5.Min}
      have \code{NA}s in them.}
    \item{\code{Parameter.2.Min}}{see \code{Parameter.1.Min}}
    \item{\code{Parameter.3.Min}}{see \code{Parameter.1.Min}}
    \item{\code{Parameter.4.Min}}{see \code{Parameter.1.Min}}
    \item{\code{Parameter.5.Min}}{see \code{Parameter.1.Min}}
    \item{\code{Parameter.1.Max}}{the columns labeled \code{Parameter.1.Max},
      \code{Parameter.2.Max}, \ldots, \cr
      \code{Parameter.5.Max} are character
      vectors containing the maximum values that can be assumed by the
      distribution parameters (see the column labeled \bold{Parameter Range(s)}
      in the table below).

      The reason these are character vectors instead of numeric vectors is
      because some parameters have an upper bound that depends on the value
      of another parameter.  For example, the parameter \code{min} for a
      \link{Uniform} distribution is bounded above by the value of the
      parameter \code{max}.

      If a distribution has \eqn{n} parameters and \eqn{n < 5}, then the
      columns labeled \code{Parameter.n+1.Max}, \ldots, \code{Parameter.5.Max}
      have the missing value code (\code{NA}).  For example, the \link{Normal}
      distribution has only two parameters associated with it (\code{mean}
      and \code{sd}) so the fields in \cr
      \code{Parameter.3.Max}, \code{Parameter.4.Max}, and \code{Parameter.5.Max}
      have \code{NA}s in them.}
    \item{\code{Parameter.2.Max}}{see \code{Parameter.1.Max}}
    \item{\code{Parameter.3.Max}}{see \code{Parameter.1.Max}}
    \item{\code{Parameter.4.Max}}{see \code{Parameter.1.Max}}
    \item{\code{Parameter.5.Max}}{see \code{Parameter.1.Max}}
  }
}
\source{
  The \pkg{EnvStats} package.
}
\references{
  Millard, S.P. (2013). \emph{EnvStats: An R Package for Environmental Statistics}.
  Springer, New York.  \url{https://link.springer.com/book/10.1007/978-1-4614-8456-1}.
}
\keyword{datasets}
