\name{EnvStats-internal}
\alias{StatNText}
\alias{StatMeanSDText}
\alias{StatMedianIQRText}
\alias{StatTestText}

\title{Internal EnvStats Objects}
\description{
  Internal EnvStats objects
}
\details{
  These are not to be called by the user.  They have been exported to allow advanced users to 
  see their structure.
}
\keyword{ internal }
