% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EnvStats-package}
\alias{EnvStats-package}
\title{EnvStats: Package for Environmental Statistics}
\description{
Graphical and statistical analyses of environmental data, with a focus on
chemical concentrations and physical parameters encountered in mandated
environmental monitoring.
}
\keyword{internal}
