% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.02d.Ex.4.mg.per.kg.vec}
\alias{EPA.02d.Ex.4.mg.per.kg.vec}
\title{Concentrations in Exhibit 4 of 2002d USEPA Guidance Document}
\usage{
data(EPA.02d.Ex.4.mg.per.kg.vec)
}
\description{
Concentrations (mg/kg) from an exposure unit.
}
\docType{data}
\format{
  a numeric vector of concentrations (mg/kg)
}
\source{
  USEPA. (2002d).  \emph{Calculating Upper Confidence Limits for Exposure Point Concentrations at
  Hazardous Waste Sites}.  OSWER 9285.6-10, December 2002.  Office of Emergency and Remedial Response,
  U.S. Environmental Protection Agency, Washington, D.C., p. 11.
}
\keyword{datasets}
