% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.09.Ex.11.1.arsenic.df}
\alias{EPA.09.Ex.11.1.arsenic.df}
\title{Arsenic Concentrations from Example 11-1 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.11.1.arsenic.df
}
\description{
Arsenic concentrations (ppb) at six wells (four observations per well).
}
\docType{data}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{Arsenic.ppb}}{a numeric vector of arsenic concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.11-3.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
