\name{EPA.09.Ex.12.4.naphthalene.df}
\alias{EPA.09.Ex.12.4.naphthalene.df}
\docType{data}
\title{
  Naphthalene Concentrations from Example 12-4 of 2009 USEPA Guidance Document
}
\description{
  Naphthalene concentrations (ppb) at five background wells (five quarterly measures at each well). 
}
\usage{EPA.09.Ex.12.4.naphthalene.df}
\format{
  A data frame with 25 observations on the following 3 variables.
  \describe{
    \item{\code{Quarter}}{a numeric vector indicating the quarter the sample was taken}
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Naphthalene.ppb}}{a numeric vector of naphthalene concentrations (ppb)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.  
  U.S. Environmental Protection Agency, Washington, D.C. p.12-12.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C. 
}
\keyword{datasets}
