% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.09.Ex.14.1.manganese.df}
\alias{EPA.09.Ex.14.1.manganese.df}
\title{Manganese Concentrations from Example 14-1 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.14.1.manganese.df
}
\description{
Manganese concentrations (ppm) at four background wells (eight quarterly measures at each well).
}
\docType{data}
\format{
  A data frame with 32 observations on the following 3 variables.
  \describe{
    \item{\code{Quarter}}{a numeric vector indicating the quarter the sample was taken}
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Manganese.ppm}}{a numeric vector of manganese concentrations (ppm)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.14-5.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
