% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Ex.14.3.alkalinity.df}
\alias{EPA.09.Ex.14.3.alkalinity.df}
\title{Alkalinity Measures from Example 14-3 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.14.3.alkalinity.df
}
\description{
Alkalinity measures (mg/L) collected from leachate at a solid waste landfill during a four and a half year period.
}
\docType{data}
\format{
  A data frame with 54 observations on the following 2 variables.
  \describe{
    \item{\code{Date}}{a Date object indicating the date of collection}
    \item{\code{Alkalinity.mg.per.L}}{a numeric vector of alkalinity measures (mg/L)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.14-14.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
