\name{EPA.09.Ex.16.2.benzene.df}
\alias{EPA.09.Ex.16.2.benzene.df}
\docType{data}
\title{
  Benzene Concentrations from Example 16-2 of 2009 USEPA Guidance Document
}
\description{
  Benzene concentrations (ppb) at one background and one downgradient well (eight monthly measures at each well). 
}
\usage{EPA.09.Ex.16.2.benzene.df}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well.type}}{a factor indicating the well type (background vs. downgradient)}
    \item{\code{Benzene.ppb}}{a numeric vector of benzene concentrations (ppb)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.  
  U.S. Environmental Protection Agency, Washington, D.C. p.16-9.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C. 
}
\keyword{datasets}
