% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Ex.17.5.chloride.df}
\alias{EPA.09.Ex.17.5.chloride.df}
\title{Chloride Concentrations from Example 17-5 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.17.5.chloride.df
}
\description{
Chloride concentrations (ppm) collected over a five-year period at a solid waste landfill.
}
\docType{data}
\format{
  A data frame with 19 observations on the following 4 variables.
  \describe{
    \item{\code{Date}}{a Date object indicating the date of collection}
    \item{\code{Chloride.ppm}}{a numeric vector of chloride concentrations (ppm)}
    \item{\code{Elapsed.Days}}{a numeric vector indicating the number of days since January 1, 2002}
    \item{\code{Residuals}}{a numeric vector of residuals from a linear regression trend fit}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.17-26.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
