% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.09.Ex.18.2.chrysene.df}
\alias{EPA.09.Ex.18.2.chrysene.df}
\title{Chrysene Concentrations from Example 18-2 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.18.2.chrysene.df
}
\description{
Chrysene concentrations (ppb) at two background wells and one compliance well (four monthly measures at each well).
}
\docType{data}
\format{
  A data frame with 12 observations on the following 4 variables.
  \describe{
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Well.type}}{a factor indicating the well type (background vs. compliance)}
    \item{\code{Chrysene.ppb}}{a numeric vector of chrysene concentrations (ppb)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.18-15.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
