% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.09.Ex.19.2.chloride.df}
\alias{EPA.09.Ex.19.2.chloride.df}
\title{Chloride Concentrations from Example 19-2 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.19.2.chloride.df
}
\description{
Chloride concentrations (mg/L) at 10 compliance wells at a solid waste landfill.
  One year of quarterly measures at each well.
}
\docType{data}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Chloride.mg.per.l}}{a numeric vector of chloride concentrations (mg/L)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.19-19.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
