\name{EPA.09.Ex.20.1.nickel.df}
\alias{EPA.09.Ex.20.1.nickel.df}
\docType{data}
\title{
  Nickel Concentrations from Example 20-1 of 2009 USEPA Guidance Document
}
\description{
  Nickel concentrations (ppb) at a single well.  Eight monthly measures during 
  the background period and eight monthly measures during the compliance period. 
}
\usage{EPA.09.Ex.20.1.nickel.df}
\format{
  A data frame with 16 observations on the following 4 variables.
  \describe{
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Year}}{a factor indicating the year of collection}
    \item{\code{Period}}{a factor indicating the period (baseline vs. compliance)}
    \item{\code{Nickel.ppb}}{a numeric vector of nickel concentrations (ppb)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.  
  U.S. Environmental Protection Agency, Washington, D.C. p.20-4.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C. 
}
\keyword{datasets}
