\name{EPA.09.Ex.21.5.beryllium.df}
\alias{EPA.09.Ex.21.5.beryllium.df}
\docType{data}
\title{
  Beryllium Concentrations from Example 21-5 of 2009 USEPA Guidance Document
}
\description{
  Beryllium concentrations (ppb) at one well (four years of quarterly measures). 
}
\usage{data(EPA.09.Ex.21.5.beryllium.df)}
\format{
  A data frame with 16 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{a factor indicating the year of collection}
    \item{\code{Quarter}}{a factor indicating the quarter of collection}
    \item{\code{Beryllium.ppb}}{a numeric vector of beryllium concentrations (ppb)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.  
  U.S. Environmental Protection Agency, Washington, D.C. p.21-18.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C. 
}
\keyword{datasets}
