\name{EPA.09.Ex.21.7.TCE.df}
\alias{EPA.09.Ex.21.7.TCE.df}
\docType{data}
\title{
  Trichloroethylene Concentrations from Example 21-7 of 2009 USEPA Guidance Document
}
\description{
  Trichloroethylene (TCE) concentrations (ppb) at a site undergoing remediation. 
}
\usage{EPA.09.Ex.21.7.TCE.df}
\format{
  A data frame with 10 observations on the following 2 variables.
  \describe{
    \item{\code{Month}}{a numeric vector indicating the month of collection}
    \item{\code{TCE.ppb}}{a numeric vector of TCE concentrations (ppb)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.  
  U.S. Environmental Protection Agency, Washington, D.C. p.21-26.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C. 
}
\keyword{datasets}
