\name{EPA.09.Ex.22.2.Specific.Conductance.df}
\alias{EPA.09.Ex.22.2.Specific.Conductance.df}
\docType{data}
\title{
  Specific Conductance from Example 22-2 of 2009 USEPA Guidance Document
}
\description{
  Specific conductance (\eqn{\mu}mho) collected over several years at two wells at a hazardous waste facility. 
}
\usage{EPA.09.Ex.22.2.Specific.Conductance.df}
\format{
  A data frame with 43 observations on the following 3 variables.
  \describe{
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Date}}{a Date object indicating the date of collection}
    \item{\code{Specific.Conductance.umho}}{a numeric vector of specific conductance (\eqn{\mu}mho)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.  
  U.S. Environmental Protection Agency, Washington, D.C. p.22-11.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C. 
}
\keyword{datasets}
