% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{EPA.09.Ex.6.3.sulfate.df}
\alias{EPA.09.Ex.6.3.sulfate.df}
\title{Sulfate Concentrations from Example 6-3 of 2009 USEPA Guidance Document}
\usage{
EPA.09.Ex.6.3.sulfate.df
}
\description{
Sulfate concentrations (ppm) at two background wells (five quarterly measures at each well).
}
\docType{data}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Month}}{a numeric vector indicating the month the observations was taken}
    \item{\code{Year}}{a numeric vector indicating the year the observation was taken}
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Sulfate.ppm}}{a numeric vector of sulfate concentrations (ppm)}
  }
}
\source{
  USEPA. (2009).  \emph{Statistical Analysis of Groundwater Monitoring Data at RCRA Facilities, Unified Guidance}.
  EPA 530/R-09-007, March 2009.  Office of Resource Conservation and Recovery Program Implementation and Information Division.
  U.S. Environmental Protection Agency, Washington, D.C. p.6-20.

  USEPA. (2010).  \emph{Errata Sheet - March 2009 Unified Guidance}.
  EPA 530/R-09-007a, August 9, 2010.  Office of Resource Conservation and Recovery, Program Information and Implementation Division.
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
