\name{EPA.89b.aldicarb2.df}
\alias{EPA.89b.aldicarb2.df}
\alias{Aldicarb}
\docType{data}
\title{
  Aldicarb Concentrations from 1989 USEPA Guidance Document
}
\description{
  Aldicarb concentrations (ppm) at three compliance wells (four monthly samples at each well). 
}
\usage{EPA.89b.aldicarb2.df}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{Aldicarb}}{Aldicarb concentrations (ppm)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1989b). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities, Interim Final Guidance}. 
  EPA/530-SW-89-026. Office of Solid Waste, U.S. Environmental Protection Agency, 
  Washington, D.C. p.6-13.
}
\keyword{datasets}
