\name{EPA.89b.chlordane1.df}
\alias{EPA.89b.chlordane1.df}
\alias{Chlordane}
\docType{data}
\title{
  Chlordane Concentrations from 1989 USEPA Guidance Document
}
\description{
  Chlordane concentrations (ppm) in 24 water samples.  Two possible phases:  dissolved 
  (18 observations) and immiscible (6 observations).
}
\usage{EPA.89b.chlordane1.df}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Chlordane}}{Chlordane concentrations (ppm)}
    \item{\code{Phase}}{a factor indicating the phase (dissolved vs. immiscible)}
  }
}
\source{
  USEPA. (1989b). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities, Interim Final Guidance}. EPA/530-SW-89-026. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C. p.4-8.
}
\keyword{datasets}
