\name{EPA.89b.edb.df}
\alias{EPA.89b.edb.df}
\docType{data}
\title{
  EDB Concentrations from 1989 USEPA Guidance Document
}
\description{
 EDB concentrations (ppb) at three compliance wells (four monthly samples at each well). 
}
\usage{EPA.89b.edb.df}
\format{
  A data frame with 12 observations on the following 3 variables.
  \describe{
    \item{\code{EDB}}{EDB concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1989b). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities, Interim Final Guidance}. EPA/530-SW-89-026. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C. p.6-6.
}
\keyword{datasets}
