\name{EPA.89b.toc.vec}
\alias{EPA.89b.toc.vec}
\docType{data}
\title{
  Total Organic Carbon Concentrations from 1989 USEPA Guidance Document
}
\description{
  Numeric vector containing total organic carbon (TOC) concentrations (mg/L).
}
\usage{EPA.89b.toc.vec}
\format{
  A numeric vector with 19 elements containing TOC concentrations (mg/L).
}
\source{
  USEPA. (1989b). 
  \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities, Interim Final Guidance}. 
  EPA/530-SW-89-026. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C. p.8-13.
}
\keyword{datasets}
