\name{EPA.92c.arsenic3.df}
\alias{EPA.92c.arsenic3.df}
\docType{data}
\title{
  Arsenic Concentrations from 1992 USEPA Guidance Document
}
\description{
  Arsenic concentrations at one background and one compliance monitoring well.  Three 
  years of observations for background well, two years of observations for 
  compliance well, four samples per year for each well. 
}
\usage{EPA.92c.arsenic3.df}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{Arsenic}}{a numeric vector of arsenic concentrations}
    \item{\code{Year}}{a factor indicating the year of collection}
    \item{\code{Well.type}}{a factor indicating the well type (background vs. compliance)}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
