% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.92c.benzene1.df}
\alias{EPA.92c.benzene1.df}
\title{Benzene Concentrations from 1992 USEPA Guidance Document}
\usage{
EPA.92c.benzene1.df
}
\description{
Benzene concentrations (ppb) at six background wells
  (six monthly samples for each well).  Nondetects reported as \code{<2}.
}
\docType{data}
\format{
  A data frame with 36 observations on the following 5 variables.
  \describe{
    \item{\code{Benzene.orig}}{a character vector of original benzene concentrations (ppb)}
    \item{\code{Benzene}}{a numeric vector of benzene concentrations with \code{<2} coded as \code{2}}
    \item{\code{Censored}}{a logical vector indicating which observations are censored}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.  p.36.
}
\keyword{datasets}
