\name{EPA.92c.ccl4.df}
\alias{EPA.92c.ccl4.df}
\docType{data}
\title{
   Carbon Tetrachloride Concentrations from 1992 USEPA Guidance Document
}
\description{
  Carbon tetrachloride (CCL4) concentrations (ppb) at five wells (four monthly samples at each well). 
}
\usage{EPA.92c.ccl4.df}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{CCL4}}{a numeric vector of carbon tetrachloride concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.  p.80.
}
\keyword{datasets}
