\name{EPA.92c.chrysene.df}
\alias{EPA.92c.chrysene.df}
\docType{data}
\title{
  Chrysene Concentrations from 1992 USEPA Guidance Document
}
\description{
  Chrysene concentrations (ppb) at five compliance wells (four monthly samples for each well). 
}
\usage{EPA.92c.chrysene.df}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{Chrysene}}{a numeric vector of chrysene concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.  p.52.
}
\keyword{datasets}
