\name{EPA.92c.copper1.df}
\alias{EPA.92c.copper1.df}
\docType{data}
\title{
  Copper Concentrations from 1992 USEPA Guidance Document
}
\description{
  Copper concentrations (ppb) at two background and one compliance wells (six monthly samples for each well). 
}
\usage{EPA.92c.copper1.df}
\format{
  A data frame with 18 observations on the following 4 variables.
  \describe{
    \item{\code{Copper}}{a numeric vector of copper concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
    \item{\code{Well.type}}{a factor indicating the well type (background vs. compliance)}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.  p.47.
}
\keyword{datasets}
