\name{EPA.92c.nickel1.df}
\alias{EPA.92c.nickel1.df}
\docType{data}
\title{
  Nickel Concentrations from 1992 USEPA Guidance Document
}
\description{
  Nickel concentrations (ppb) at four monitoring wells (five monthly samples for each well). 
}
\usage{EPA.92c.nickel1.df}
\format{
  A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{Nickel}}{a numeric vector of nickel concentrations (ppb)}
    \item{\code{Month}}{a factor indicating the month of collection}
    \item{\code{Well}}{a factor indicating the well number}
  }
}
\source{
  USEPA. (1992c). \emph{Statistical Analysis of Ground-Water Monitoring Data at RCRA Facilities: Addendum to Interim Final Guidance}. Office of Solid Waste, U.S. Environmental Protection Agency, Washington, D.C.  p.7.
}
\keyword{datasets}
