\name{EPA.92d.chromium.df}
\alias{EPA.92d.chromium.df}
\docType{data}
\title{
  Chromium Concentrations from 1992 USEPA Guidance Document
}
\description{
  Chromium concentrations (mg/kg) in soil samples collected randomly over a Superfund site.
}
\usage{EPA.92d.chromium.df}
\format{
  A data frame with 15 observations on the following variable.
  \describe{
    \item{\code{Cr}}{a numeric vector of chromium concentrations (mg/kg)}
  }
}
\source{
  USEPA. (1992d). \emph{Supplemental Guidance to RAGS: Calculating the Concentration Term}.  
  Publication 9285.7-081, May 1992. Intermittent Bulletin, Volume 1, Number 1.  
  Office of Emergency and Remedial Response, Hazardous Site Evaluation Division, OS-230.  
  Office of Solid Waste and Emergency Response, 
  U.S. Environmental Protection Agency, Washington, D.C.
}
\keyword{datasets}
