% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{EPA.97.cadmium.111.df}
\alias{EPA.97.cadmium.111.df}
\title{Calibration Data for Cadmium at Mass 111}
\usage{
EPA.97.cadmium.111.df
}
\description{
Calibration data for cadmium at mass 111 (ng/L; method 1638 ICPMS) that appeared in
  Gibbons et al. (1997b) and were provided to them by the U.S. EPA.
}
\docType{data}
\format{
  A data frame with 35 observations on the following 2 variables.
  \describe{
    \item{\code{Cadmium}}{Observed concentation of cadmium (ng/L)}
    \item{\code{Spike}}{\dQuote{True} concentration of cadmium taken from a standard (ng/L)}
  }
}
\source{
  Gibbons, R.D., D.E. Coleman, and R.F. Maddalone. (1997b). Response to Comment on "An Alternative Minimum Level Definition for Analytical Quantification".
  \emph{Environmental Science and Technology}, \bold{31}(12), 3729--3731.
}
\keyword{datasets}
