% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{FcnsByCatDataTrans}
\alias{FcnsByCatDataTrans}
\title{EnvStats Functions for Data Transformations}
\description{
The \pkg{EnvStats} functions listed below are useful for deciding on data transformations.
}
\alias{Data Transformations}
\details{
  \tabular{ll}{
  \emph{Function Name}                    \tab \emph{Description} \cr
  \code{\link{boxcox}}                    \tab Compute values of an objective for Box-Cox transformations, or \cr
                                          \tab compute optimal transformation based on raw observations \cr
                                          \tab or residuals from a linear model. \cr
  \code{\link{boxcoxTransform}}           \tab Apply a Box-Cox Power transformation to a set of data. \cr
  \code{\link{plot.boxcox}}               \tab Plotting method for an object of class \code{\link[=boxcox.object]{"boxcox"}}. \cr
  \code{\link{plot.boxcoxLm}}             \tab Plotting method for an object of class \code{\link[=boxcoxLm.object]{"boxcoxLm"}}. \cr
  \code{\link{print.boxcox}}              \tab Printing method for an object of class \code{\link[=boxcox.object]{"boxcox"}}. \cr
  \code{\link{print.boxcoxLm}}            \tab Printing method for an object of class \code{\link[=boxcoxLm.object]{"boxcoxLm"}}. \cr
  }
}
\keyword{ package }
