% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{FcnsByCatMCandRisk}
\alias{FcnsByCatMCandRisk}
\title{EnvStats Functions for Monte Carlo Simulation and Risk Assessment}
\description{
The \pkg{EnvStats} functions listed below are useful for performing Monte Carlo simulations
  and risk assessment.
}
\alias{Monte Carlo Simulation and Risk Assessment}
\details{
  \tabular{ll}{
  \emph{Function Name}                    \tab \emph{Description} \cr
  \link{Empirical}                        \tab Empirical distribution based on a set of observations. \cr
  \code{\link{simulateVector}}            \tab Simulate a vector of random numbers from a specified theoretical \cr
                                          \tab probability distribution or empirical probability distribution \cr
                                          \tab using either Latin hypercube sampling or simple random sampling. \cr
  \code{\link{simulateMvMatrix}}          \tab Simulate a multivariate matrix of random numbers from specified \cr
                                          \tab theoretical probability distributions and/or empirical probability \cr
                                          \tab distributions based on a specified rank correlation matrix, using \cr
                                          \tab either Latin hypercube sampling or simple random sampling. \cr
  }
}
\keyword{ package }
