% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{FcnsByCatPower}
\alias{FcnsByCatPower}
\title{EnvStats Functions for Power and Sample Size Calculations}
\description{
The \pkg{EnvStats} functions listed below are useful for power and sample size calculations.
}
\details{
This help page has been shortened to keep function help focused on usage,
arguments, return values, and examples. Extended method details, formulas,
and background material are available in \code{vignette("extended-function-details", package = "EnvStats")}, section \code{FcnsByCatPower}.
}
\alias{Power and Sample Size Calculations}
\alias{Power and Sample Size}
\keyword{ package }
