\name{FcnsByCatPrintPlot}
\alias{FcnsByCatPrintPlot}
\alias{Printing and Plotting Methods}
\title{
  EnvStats Functions for Printing and Plotting Objects of Various S3 Classes
}
\description{
  The \pkg{EnvStats} functions listed below are printing and plotting methods for various S3 classes.
}
\details{
  \bold{Printing Methods}

  \tabular{ll}{
  \emph{Function Name}                    \tab \emph{Description} \cr
  \code{\link{print.boxcox}}              \tab Print an object that inherits from class \code{\link[=boxcox.object]{"boxcox"}}. \cr
  \code{\link{print.boxcoxCensored}}      \tab Print an object that inherits from class \cr
                                          \tab \code{\link[=boxcoxCensored.object]{"boxcoxCensored"}}. \cr
  \code{\link{print.boxcoxLm}}            \tab Print an object that inherits from class \code{\link[=boxcoxLm.object]{"boxcoxLm"}}. \cr
                                          \tab \cr
  \code{\link{print.estimate}}            \tab Print an object that inherits from class \code{\link[=estimate.object]{"estimate"}}. \cr
  \code{\link{print.estimateCensored}}    \tab Print an object that inherits from class \cr
                                          \tab \code{\link[=estimateCensored.object]{"estimateCensored"}}. \cr
                                          \tab \cr
  \code{\link{print.gof}}                 \tab Print an object that inherits from class \code{\link[=gof.object]{"gof"}}. \cr
  \code{\link{print.gofCensored}}         \tab Print an object that inherits from class \code{\link[=gofCensored.object]{"gofCensored"}}. \cr
  \code{\link{print.gofGroup}}            \tab Print an object that inherits from class \code{\link[=gofGroup.object]{"gofGroup"}}. \cr
  \code{\link{print.gofTwoSample}}        \tab Print an object that inherits from class \cr
                                          \tab \code{\link[=gofTwoSample.object]{"gofTwoSample"}}. \cr
                                          \tab \cr
  \code{\link{print.htest}}               \tab Print an object that inherits from class \code{\link[=htest.object]{"htest"}}. \cr
  \code{\link{print.htestCensored}}       \tab Print an object that inherits from class \cr
                                          \tab \code{\link[=htestCensored.object]{"htestCensored"}}. \cr
  \code{\link{print.permutationTest}}     \tab Print an object that inherits from class \cr
                                          \tab \code{\link[=permutationTest.object]{"permutationTest"}}. \cr
                                          \tab \cr
  \code{\link{print.summaryStats}}        \tab Print an object that inherits from class \cr
                                          \tab \code{\link[=summaryStats.object]{"summaryStats"}}. \cr
  }


  \bold{Plotting Methods}

  \tabular{ll}{
  \emph{Function Name}                    \tab \emph{Description} \cr
  \code{\link{plot.boxcox}}               \tab Plot an object that inherits from class \code{\link[=boxcox.object]{"boxcox"}}. \cr
  \code{\link{plot.boxcoxCensored}}       \tab Plot an object that inherits from class \code{\link[=boxcoxCensored.object]{"boxcoxCensored"}}. \cr
  \code{\link{plot.boxcoxLm}}             \tab Plot an object that inherits from class \code{\link[=boxcoxLm.object]{"boxcoxLm"}}. \cr
                                          \tab \cr
  \code{\link{plot.gof}}                  \tab Plot an object that inherits from class \code{\link[=gof.object]{"gof"}}. \cr
  \code{\link{plot.gofCensored}}          \tab Plot an object that inherits from class \code{\link[=gofCensored.object]{"gofCensored"}}. \cr
  \code{\link{plot.gofGroup}}             \tab Plot an object that inherits from class \code{\link[=gofGroup.object]{"gofGroup"}}. \cr
  \code{\link{plot.gofTwoSample}}         \tab Plot an object that inherits from class \code{\link[=gofTwoSample.object]{"gofTwoSample"}}. \cr
                                          \tab \cr
  \code{\link{plot.permutationTest}}      \tab Plot an object that inherits from class \code{\link[=permutationTest.object]{"permutationTest"}}. \cr
  }
}
\keyword{ package }


