\name{FcnsByCatTolInts}
\alias{FcnsByCatTolInts}
\alias{Tolerance Intervals}
\title{
  EnvStats Functions for Tolerance Intervals
}
\description{
  The \pkg{EnvStats} functions listed below are useful for computing tolerance intervals.  
  See \link[=FcnsByCatPower]{Power and Sample Size} for a list of functions useful for 
  computing power and sample size for a design based on a tolerance interval width.
}
\details{

  \tabular{ll}{
  \emph{Function Name}                    \tab \emph{Description} \cr
  \code{\link{tolIntGamma}},              \tab Tolerance interval for a Gamma distribution. \cr
  \code{\link{tolIntGammaAlt}}            \tab \cr
                                          \tab \cr
  \code{\link{tolIntLnorm}},              \tab Tolerance interval for a lognormal distribution. \cr
  \code{\link{tolIntLnormAlt}}            \tab \cr
                                          \tab \cr
  \code{\link{tolIntNorm}}                \tab Tolerance interval for a Normal (Gaussian) distribution. \cr
  \code{\link{tolIntNormK}}               \tab Compute the constant \eqn{K} for a Normal (Gaussian) \cr
                                          \tab tolerance interval. \cr
                                          \tab \cr
  \code{\link{tolIntNpar}}                \tab Nonparametric tolerance interval. \cr
                                          \tab \cr
  \code{\link{tolIntPois}}                \tab Tolerance interval for a Poisson distribution. \cr
  } 
}
\keyword{ package }
