% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{Gibbons.et.al.09.Alkilinity.vec}
\alias{Gibbons.et.al.09.Alkilinity.vec}
\title{Alkilinity Data from Gibbons et al. (2009)}
\usage{
data(Gibbons.et.al.09.Alkilinity.vec)
}
\description{
Alkilinity concentrations (mg/L) in groundwater.
}
\docType{data}
\format{
  A numeric vector with 27 elements.
}
\source{
  Gibbons, R.D., D.K. Bhaumik, and S. Aryal. (2009).
  \emph{Statistical Methods for Groundwater Monitoring}. Second Edition.
  John Wiley & Sons, Hoboken.  Table 5.5, p. 107.
}
\keyword{datasets}
