\name{Gibbons.et.al.09.Alkilinity.vec}
\alias{Gibbons.et.al.09.Alkilinity.vec}
\docType{data}
\title{
  Alkilinity Data from Gibbons et al. (2009)
}
\description{
  Alkilinity concentrations (mg/L) in groundwater.
}
\usage{data(Gibbons.et.al.09.Alkilinity.vec)}
\format{
  A numeric vector with 27 elements.
}
\source{
  Gibbons, R.D., D.K. Bhaumik, and S. Aryal. (2009). 
  \emph{Statistical Methods for Groundwater Monitoring}. Second Edition.  
  John Wiley & Sons, Hoboken.  Table 5.5, p. 107.
}
\keyword{datasets}
