\name{Gibbons.et.al.09.Vinyl.Chloride.vec}
\alias{Gibbons.et.al.09.Vinyl.Chloride.vec}
\docType{data}
\title{
  Vinyl Chloride Data from Gibbons et al. (2009)
}
\description{
  Vinyl chloride concentrations (\eqn{mu}g/L) in groundwater from upgradient 
  background monitoring wells.
}
\usage{data(Gibbons.et.al.09.Vinyl.Chloride.vec)}
\format{
  A numeric vector with 34 elements.
}
\source{
  Gibbons, R.D., D.K. Bhaumik, and S. Aryal. (2009). 
  \emph{Statistical Methods for Groundwater Monitoring}. Second Edition.  
  John Wiley & Sons, Hoboken.  Table 4.3, p. 87.
}
\keyword{datasets}
