% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{Graham.et.al.75.etu.df}
\alias{Graham.et.al.75.etu.df}
\title{Ethylene Thiourea Dose-Response Data}
\usage{
Graham.et.al.75.etu.df
}
\description{
These data are the results of an experiment in which different groups of rats were
  exposed to different concentration levels of ethylene thiourea (ETU), which is a
  decomposition product of a certain class of fungicides that can be found in treated
  foods (Graham et al., 1975; Rodricks, 1992, p.133).  In this experiment, the outcome of
  concern was the number of rats that developed thyroid tumors.
}
\docType{data}
\format{
  A data frame with 6 observations on the following 4 variables.
  \describe{
    \item{\code{dose}}{a numeric vector of dose (ppm/day) of ETU.}
    \item{\code{tumors}}{a numeric vector indicating number of rats that developed thyroid tumors.}
    \item{\code{n}}{a numeric vector indicating the number of rats in the dose group.}
    \item{\code{proportion}}{a numeric vector indicating proportion of rats that developed thyroid tumors.}
  }
}
\source{
  Graham, S.L., K.J. Davis, W.H. Hansen, and C.H. Graham. (1975).
  Effects of Prolonged Ethylene Thiourea Ingestion on the Thyroid of the Rat.
  \emph{Food and Cosmetics Toxicology}, \bold{13}(5), 493--499.
}
\references{
  Rodricks, J.V. (1992). \emph{Calculated Risks: The Toxicity and Human Health Risks of Chemicals in Our Environment}.
  Cambridge University Press, New York, p.133.
}
\keyword{datasets}
