\name{Helsel.Hirsch.02.Mayfly.df}
\alias{Helsel.Hirsch.02.Mayfly.df}
\docType{data}
\title{
  Paired Counts of Mayfly Nymphs Above and Below Industrial Outfalls
}
\description{
  Counts of mayfly nymphs at low flow in 12 small streams.
  In each stream, counts were recorded above and below industrial outfalls.
}
\usage{data(Helsel.Hirsch.02.Mayfly.df)}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{Mayfly.Count}}{Number of mayfly nymphs counted}
    \item{\code{Stream}}{a factor indicating the stream number}
    \item{\code{Location}}{a factor indicating the location of the count (above vs. below)}
  }
}
\source{
  Helsel, D.R., and R.M. Hirsch. (2002). \emph{Statistical Methods in Water Resources Research}.
  Techniques of Water Resources Investigations, Book 4, Chapter A3.  U.S. Geological Survey,
  139--140.  \url{https://pubs.usgs.gov/tm/04/a03/tm4a3.pdf}.
}
\keyword{datasets}
