% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{Lin.Evans.80.df}
\alias{Lin.Evans.80.df}
\title{Fecal Coliform Data from the Illinois River}
\usage{
Lin.Evans.80.df
}
\description{
Lin and Evans (1980) reported fecal coliform measures (organisms per 100 ml) from the
  Illinois River taken between 1971 and 1976.  The object \code{Lin.Evans.80.df} is a
  small subset of these data that were reported by Helsel and Hirsch (1992, p.162).
}
\docType{data}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{Fecal.Coliform}}{a numeric vector of fecal coliform measure (organisms per 100 ml).}
    \item{\code{Season}}{an ordered factor indicating the season of collection}
  }
}
\source{
  Helsel, D.R., and R.M. Hirsch. (1992). \emph{Statistical Methods in Water Resources Research}.
  Elsevier, New York, NY, p.162.
}
\references{
  Lin, S.D., and R.L. Evans. (1980). \emph{Coliforms and fecal streptococcus in the Illinois River at Peoria, 1971-1976}.
  Illinois State Water Survey Report of Investigations No. 93. Urbana, IL, 28pp.
}
\keyword{datasets}
