\name{newsEnvStats}
\alias{newsEnvStats}
\title{
  Show the EnvStats NEWS File
}
\description{
  Show the NEWS file of the \pkg{EnvStats} package. 
}
\usage{
  newsEnvStats()
}
\details{
  The function \code{newsEnvStats} displays the contents of the \pkg{EnvStats} NEWS file in a 
  separate text window.  You can also access the NEWS file with the command 
  \code{news(package="EnvStats")}, which returns the contents of the file to the \R command 
  window.
}
\value{
  None.
}
\author{
  Steven P. Millard (\email{EnvStats@ProbStatInfo.com})
}
\seealso{
  \code{\link{news}}.
}
\keyword{ package }
