% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0package.R
\name{NIOSH.89.air.lead.vec}
\alias{NIOSH.89.air.lead.vec}
\title{NIOSH Air Lead Levels Data}
\usage{
NIOSH.89.air.lead.vec
}
\description{
Air lead levels collected by the National Institute for Occupational Safety and Health
  (NIOSH) at 15 different areas within the Alma American Labs, Fairply, CO,
  for health hazard evaluation (HETA 89-052) on Februay 23, 1989.
}
\docType{data}
\format{
  A numeric vector with 15 elements containing air lead concentrations (\eqn{\mu g/m^3}).
}
\source{
  Krishnamoorthy, K., T. Matthew, and G. Ramachandran. (2006).
  Generalized P-Values and Confidence Intervals:  A Novel Approach for Analyzing
  Lognormally Distributed Exposure Data.
  \emph{Journal of Occupational and Environmental Hygiene}, \bold{3}, 642--650.
}
\references{
  Zou, G.Y., C.Y. Huo, and J. Taleban. (2009).  Simple Confidence Intervals for
  Lognormal Means and their Differences with Environmental Applications.
  \emph{Environmetrics}, \bold{20}, 172--180.
}
\keyword{datasets}
